Changed crosshairs to red.  Updated some variable names.
When I wanted to change the crosshairs color, I discovered that
it's also useful to add a context.beginPath() method call before
starting a new stroke.  Otherwise, the axes were also getting
redrawn as red instead of remaining black.

The variable names were updated for readability, and searchability
in the cases of i->ii and j->jj.